Rename metadata subpackage to dsse
* Moves DSSE Envelope from metadata.py to dsse.py.
* Moves base64 functions from util.py to _internal/utils.py
* Minor documentation fixes
